+++ libs/qsynedit/qsynedit/syntaxer/cpp.h
@@ -159,6 +159,8 @@ private:
     int mLeftBraces;
     int mRightBraces;
 
+    QString mRawStringInitialDCharSeq;
+
     QSet<QString> mCustomTypeKeywords;
 
     PTokenAttribute mPreprocessorAttribute;